Move login attempt verifier
